+++ print_service/utils/__init__.py
@@ -13,6 +13,11 @@ from print_service.models import File
 from print_service.models import File as FileModel
 from print_service.settings import Settings, get_settings
 
+from PIL import Image
+from io import BytesIO
+from fpdf import FPDF
+
+
 
 settings: Settings = get_settings()
 
@@ -67,3 +72,23 @@ def get_file(dbsession, pin: str or list[str]):
             },
         })
     return result
+
+#преобразует файл из картинки в pdf
+def process_image(file :str,newFileName :str):
+    img=Image.open(BytesIO(file)).convert("RGB")  
+    if(img.width>img.height):
+        img=img.rotate(90,expand=True)
+    #img.save(f"{newFileName}.pdf", "PDF", quality=100)
+    pdf = FPDF()
+    pdf.add_page()
+
+    FPDF.set_left_margin(pdf,0)
+    FPDF.set_right_margin(pdf,0)
+    FPDF.set_top_margin(pdf,10)
+    
+    k=min(pdf.eph/img.height,pdf.epw/img.width)
+    wid=img.width*k
+    heig=img.height*k
+    #print(wid,heig,pdf.epw,pdf.eph, (pdf.eph-heig)/2)
+    pdf.image(img, x=(pdf.epw-wid)/2,y=(pdf.eph-heig)/2+15,h=heig, w=wid) 
+    pdf.output(f"{newFileName}.pdf")
